+++ src/controllers/index.ts
@@ -1,5 +1,6 @@
 import UserController from './UserController';
 import MumentController from './MumentController';
 import MusicController from './MusicController';
+import AuthController from './AuthController';
 
-export { UserController, MumentController, MusicController };
+export { UserController, MumentController, MusicController, AuthController };
