+++ src/routes/UserRouter.ts
@@ -7,10 +7,4 @@ const router: Router = Router();
 router.get('/my/:userId/list', UserController.getMyMumentList);
 router.get('/like/:userId/list', UserController.getLikeMumentList);
 
-// 로그인
-router.post('/auth/login', [
-    body('profileId').notEmpty().isString(), 
-    body('password').notEmpty().isString(),
-], UserController.login);
-
 export default router;
