+++ src/services/index.ts
@@ -1,5 +1,6 @@
 import UserService from './UserService';
 import MumentService from './MumentService';
 import MusicService from './MusicService';
+import AuthService from './AuthService';
 
-export { UserService, MumentService, MusicService };
+export { UserService, MumentService, MusicService, AuthService };
